LayoutTests/editing/selection/ios/no-click-event-when-adjusting-selection-handles-expected.txt

This test verifies that interacting with selection handles doesn't trigger a synthetic click. To manually run the test, select a word in the paragraph above and then adjust either selection handle by a single letter. The selection should not disappear.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS firstSelectionRect.width > 0 is true
PASS firstSelectionRect.height > 0 is true
PASS successfullyParsed is true

TEST COMPLETE
